hw13_calendar is completed

+++ hw12_13_14_15_calendar/internal/storage/event.go
@@ -9,5 +9,5 @@ type Event struct {
 	ExpirationDate           time.Time
 	Description              string
 	UserID                   string
-	SecondsUntilNotification int
+	SecondsUntilNotification int64
 }
